backend/classifier/requirements.txt
torch
transformers
Pillow
ftfy
open_clip_torch